% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash_file_with_blake3}
\alias{hash_file_with_blake3}
\title{hash a file with the blake3 algorithm}
\usage{
hash_file_with_blake3(file, memmap = TRUE, parallelize = TRUE)
}
\arguments{
\item{file}{path to file}

\item{memmap}{whether to use memory mapping}

\item{parallelize}{use parallel hash}
}
\description{
hash a file with the blake3 algorithm
}
