% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hash_file_with_blake3_direct_impl}
\alias{hash_file_with_blake3_direct_impl}
\title{Return string \code{"Hello world!"} to R.}
\usage{
hash_file_with_blake3_direct_impl(file_path)
}
\description{
Return string \code{"Hello world!"} to R.
}
